Refactoring an polish for threshold output
