+++ api/jstachio/src/main/java/io/jstach/jstachio/output/package-info.java
@@ -1,5 +1,6 @@
 /**
- * Output utility classes particularly for leveraging pre-encoding.
+ * Output utility classes particularly for leveraging pre-encoding efficiently and shared
+ * agnostic integration logic for plugging into web application frameworks.
  * @see io.jstach.jstachio.output.BufferedEncodedOutput
  * @see io.jstach.jstachio.output.ThresholdEncodedOutput
  * @apiNote As with most IO the classes in this package are not thread safe unless noted.
