+++ api/jstachio/src/main/java/module-info.java
@@ -26,6 +26,9 @@
  * </pre>
  *
  * @see io.jstach.jstachio.JStachio
+ * @uses io.jstach.jstachio.spi.JStachioExtension JStachio runtime extension SPI
+ * @uses io.jstach.jstachio.spi.TemplateProvider SPI used to find generated templates 
+ * (see {@link io.jstach.jstache.JStacheCatalog}).
  * @jstachioVersion
  */
 module io.jstach.jstachio {
